+++ mupen64plus-core/src/r4300/x86_64/gcop1_w.c
@@ -40,7 +40,7 @@ void gencvt_s_w(void)
    inc_m32rel(&instr_count[117]);
 #endif
 #ifdef INTERPRET_CVT_S_W
-   gencallinterp((unsigned long long)cached_interpreter_table.CVT_S_W, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.CVT_S_W, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_simple[dst->f.cf.fs]));
@@ -56,7 +56,7 @@ void gencvt_d_w(void)
    inc_m32rel(&instr_count[117]);
 #endif
 #ifdef INTERPRET_CVT_D_W
-   gencallinterp((unsigned long long)cached_interpreter_table.CVT_D_W, 0);
+   gencallinterp((uint64_t)cached_interpreter_table.CVT_D_W, 0);
 #else
    gencheck_cop1_unusable();
    mov_xreg64_m64rel(RAX, (unsigned long long *)(&reg_cop1_simple[dst->f.cf.fs]));
